Update controllers.py
cononical_title will set to the title if the value isn't provided.

thumbnail_poster_url will set to the full_size_poster_url if none is provided. It will be empty if no full_size_poster_url is not provided either. 
The full_size_poster_url is set in the same way.